Refactor functions into different files, tweak filenames.
